Re-introduce flake8 doc linting + doc fixes
