+++ dev-requirements/flake8.in
@@ -10,8 +10,7 @@ flake8-builtins==1.5.3           # builtin shadowing checks
 flake8-coding==1.3.2             # coding magic-comment detection
 flake8-comprehensions==3.10.0       # comprehension checks
 flake8-deprecated==1.3           # deprecated call checks
-# flake8-docstrings==1.6         # pydocstyle support
-# D401 messes with property docs too much
+flake8-docstrings==1.6         # pydocstyle support
 flake8-executable==2.1.1         # shebangs
 flake8-fixme==1.1.1              # "fix me" counter
 flake8-functions==0.0.7          # function linting
